Restructured New Code to Match Current Setup
