+++ android/app/build.gradle
@@ -30,15 +30,17 @@ android {
 }
 
 repositories {
+    google()
     flatDir{
         dirs '../capacitor-cordova-android-plugins/src/main/libs', 'libs'
     }
-    maven {
-        url "https://repo.osgeo.org/repository/release/"
-    }
+
 }
 
 dependencies {
+    implementation 'com.mapbox.mapboxsdk:mapbox-sdk-geojson:5.9.0-alpha.1'
+    implementation 'com.mapbox.mapboxsdk:mapbox-sdk-turf:5.9.0-alpha.1'
+    implementation 'com.mapbox.mapboxsdk:mapbox-sdk-core:5.9.0-alpha.1'
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation "androidx.appcompat:appcompat:$androidxAppCompatVersion"
     implementation "androidx.coordinatorlayout:coordinatorlayout:$androidxCoordinatorLayoutVersion"
@@ -50,9 +52,6 @@ dependencies {
     implementation project(':capacitor-cordova-android-plugins')
     implementation 'com.android.volley:volley:1.2.1'
     implementation 'com.google.code.gson:gson:2.9.0'
-    implementation 'org.locationtech.jts:jts-core:1.19.0'
-    implementation 'org.geotools:gt-main:24.1'
-    implementation 'org.geotools:gt-geojson:24.1'
 }
 
 apply from: 'capacitor.build.gradle'
